examples/smolagents-localmodels/requirements.txt
# Core ML libraries
torch
transformers>=4.36.0
accelerate>=0.27.0
huggingface-hub>=0.20.0

# Agent và tools
smolagents>=0.0.6
litellm>=1.0.0

# Environment và utils
python-dotenv>=0.19.0
requests>=2.31.0
numpy>=1.24.0

# Optional but recommended
sentencepiece  # Cho tokenization
protobuf  # Cho model serialization

# Web API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart # Form data
pydantic>=2.0.0 # Data validation như TypeScript

# CORS
starlette>=0.36.0
cors>=1.0.1
